backend/requirements.txt
fastapi
uvicorn
PyJWT
python-dotenv
stripe
pydantic
livekit-api
requests
cryptography
httpx
beautifulsoup4
langchain
langchain-community
langchain-core
langchain-openai
Pillow
mcp
